scripts/requirements.txt
gitdb==0.6.4
GitPython==1.0.1
lxml==3.4.4
smmap==0.9.0
Whoosh==2.7.0
